+++ crates/cli/tests/plugins_test.rs
@@ -370,6 +370,26 @@ mod plugins {
         }
 
         #[test]
+        fn supports_poetry() {
+            let sandbox = create_empty_proto_sandbox();
+
+            sandbox
+                .run_bin(|cmd| {
+                    cmd.arg("install").arg("poetry");
+                })
+                .success();
+
+            create_shim_command(sandbox.path(), "poetry")
+                .arg("--version")
+                .assert()
+                .success();
+
+            assert_snapshot!(
+                fs::read_to_string(sandbox.path().join(".proto/shims/registry.json")).unwrap()
+            );
+        }
+
+        #[test]
         fn supports_uv() {
             let sandbox = create_empty_proto_sandbox();