+++ crates/core/src/config.rs
@@ -395,6 +395,15 @@ impl ProtoConfig {
             }
         }
 
+        if !self.plugins.contains_key("poetry") && is_allowed("poetry") {
+            self.plugins.insert(
+                Id::raw("poetry"),
+                PluginLocator::Url(Box::new(UrlLocator {
+                    url: "https://github.com/moonrepo/plugins/releases/download/python_poetry_tool-v0.1.0/python_poetry_tool.wasm".into()
+                }))
+            );
+        }
+
         if !self.plugins.contains_key("python") && is_allowed("python") {
             self.plugins.insert(
                 Id::raw("python"),